Alterações na tela principal e scan
